requirements.txt
# Microservices
flask
flasgger
flask-mongoengine
pymongo
celery
# General
boto3==1.12.40 # last version does not include docevents
environs
#blas
# Testing
pytest-flask
pytest-mock
pytest-cov
# ML
pandas
#seaport
numpy
matplotlib
seaborn
